# docs/openapi/openapi-operational-milestones.yml
openapi: 3.0.3

info:
  title: ODP REST API / Operational Milestones
  description: Operational Change Milestones Management Operations for the Operational Deployment Plan Management API
  version: 1.0.0
  contact:
    name: ODP Development Team

paths:
  # Milestone CRUD Endpoints
  /operational-changes/{id}/milestones:
    get:
      summary: List all milestones for operational change (latest version, baseline context, or wave filtered)
      tags: [Milestones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Operational Change Item ID
        - name: baseline
          in: query
          required: false
          schema:
            type: string
          description: "Optional baseline ID to get milestones as they existed at baseline time"
        - name: fromWave
          in: query
          required: false
          schema:
            type: string
          description: "Optional wave ID for filtering"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'openapi-base.yml#/components/schemas/Milestone'
        '404':
          description: Operational Change not found or not captured in specified baseline
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'
        '400':
          description: Bad Request - Invalid baseline or wave ID
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'

    post:
      summary: Add milestone to operational change (creates new OC version)
      tags: [Milestones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Operational Change Item ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'openapi-base.yml#/components/schemas/MilestoneUpdateRequest'
      responses:
        '201':
          description: Milestone created and new OC version generated
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/MilestoneCreateResponse'
        '409':
          description: Version Conflict - expectedVersionId is outdated
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'
        '400':
          description: Validation Error - Invalid milestone data or wave reference
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'
        '404':
          description: Operational Change not found
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'

  /operational-changes/{id}/milestones/{milestoneKey}:
    get:
      summary: Get specific milestone (latest version, baseline context, or wave filtered)
      tags: [Milestones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Operational Change Item ID
        - name: milestoneKey
          in: path
          required: true
          schema:
            type: string
          description: Milestone Key (stable identifier preserved across versions)
        - name: baseline
          in: query
          required: false
          schema:
            type: string
          description: "Optional baseline ID to get milestone as it existed at baseline time"
        - name: fromWave
          in: query
          required: false
          schema:
            type: string
          description: "Optional wave ID for filtering"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Milestone'
        '404':
          description: Milestone or Operational Change not found, or not captured in specified baseline
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'
        '400':
          description: Bad Request - Invalid baseline or wave ID
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'

    put:
      summary: Update milestone (creates new OC version)
      tags: [Milestones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Operational Change Item ID
        - name: milestoneKey
          in: path
          required: true
          schema:
            type: string
          description: Milestone Key (stable identifier preserved across versions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'openapi-base.yml#/components/schemas/MilestoneUpdateRequest'
      responses:
        '200':
          description: Milestone updated and new OC version generated
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/MilestoneUpdateResponse'
        '409':
          description: Version Conflict - expectedVersionId is outdated
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'
        '400':
          description: Validation Error - Invalid milestone data or wave reference
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'
        '404':
          description: Milestone or Operational Change not found
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'

    delete:
      summary: Delete milestone (creates new OC version)
      tags: [Milestones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Operational Change Item ID
        - name: milestoneKey
          in: path
          required: true
          schema:
            type: string
          description: Milestone Key (stable identifier preserved across versions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersionId]
              properties:
                expectedVersionId:
                  type: string
                  description: "Version ID of the parent OperationalChange (for optimistic locking)"
      responses:
        '200':
          description: Milestone deleted and new OC version generated
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/MilestoneDeleteResponse'
        '409':
          description: Version Conflict - expectedVersionId is outdated
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'
        '404':
          description: Milestone or Operational Change not found
          content:
            application/json:
              schema:
                $ref: 'openapi-base.yml#/components/schemas/Error'